* change name GenericObject to SearchResponse; * change variable user_id to userId/userHid; * moved classes to separate files;
