+++ src/main/java/lesson7/Meta.java
@@ -1,0 +1,13 @@
+package lesson7;
+
+public class Meta {
+    private Response response;
+
+    public Response getResponse() {
+        return response;
+    }
+
+    public void setResponse(Response response) {
+        this.response = response;
+    }
+}
